+++ Priority Queue/SimplePriorityQueue.cs
@@ -186,7 +186,7 @@ namespace Priority_Queue
         /// Duplicates are allowed.
         /// O(log n)
         /// </summary>
-        public void Enqueue(T item, double priority)
+        public void Enqueue(T item, long priority)
         {
             lock(_queue)
             {
@@ -228,7 +228,7 @@ namespace Priority_Queue
         /// to update all of them, please wrap your items in a wrapper class so they can be distinguished).
         /// O(n)
         /// </summary>
-        public void UpdatePriority(T item, double priority)
+        public void UpdatePriority(T item, long priority)
         {
             lock (_queue)
             {
